wip changes for beacon handlers
