--- proto/eth/v2/beacon_chain.pb.gw.go
@@ -1,4 +1,0 @@
-//go:build ignore
-// +build ignore
-
-package ignore
