--- proto/eth/v2/beacon_chain.proto
@@ -1,29 +1,0 @@
-// Copyright 2022 Prysmatic Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-syntax = "proto3";
-
-package ethereum.eth.v2;
-
-import "proto/eth/v2/withdrawals.proto";
-
-option csharp_namespace = "Ethereum.Eth.V2";
-option go_package = "github.com/prysmaticlabs/prysm/v4/proto/eth/v2;eth";
-option java_multiple_files = true;
-option java_outer_classname = "BeaconChainProto";
-option java_package = "org.ethereum.eth.v2";
-option php_namespace = "Ethereum\\Eth\\v2";
-
-message BLSToExecutionChangesPoolResponse {
-  repeated SignedBLSToExecutionChange data = 1;
-}
